Employeee Attrition/requirements.txt
streamlit
pandas
numpy
matplotlib
seaborn
scikit-learn
Pillow
requests
